chore: move dataclasses to schema
